Did a bunch of refactoring and renaming to better structure things
