+++ Kanban.csproj
@@ -14,7 +14,6 @@
     <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="7.0.7" />
     <PackageReference Include="MudBlazor" Version="6.16.0" />
     <PackageReference Include="RecurrenceCalculator" Version="3.0.2" />
-    <PackageReference Include="Swashbuckle.AspNetCore" Version="6.5.0" />
   </ItemGroup>
 
   <ItemGroup>
