+++ Program.cs
@@ -14,8 +14,6 @@ builder.Services.AddControllers()
 				.AddNewtonsoftJson ();
 builder.Services.AddControllersWithViews();
 
-// Learn more about configuring Swagger/OpenAPI at https://aka.ms/aspnetcore/swashbuckle
-builder.Services.AddEndpointsApiExplorer ();//I think Swagger can be removed soon
 //builder.Services.AddCors ();
 builder.Services.AddHttpClient ();
