+++ Repositories/BoardRepository.cs
@@ -1,0 +1,33 @@
+﻿using Azure.Data.Tables;
+using Kanban.Contexts;
+using Microsoft.Extensions.Options;
+
+namespace Kanban.Repositories;
+
+public class BoardRepository
+{
+    private const string boards = "Boards";
+    private const string columns = "Columns";
+    private const string swimlanes = "Swimlanes";
+    private const string cards = "Cards";
+    private const string tags = "Tags";
+
+    private readonly TableServiceClient _tableServiceClient;
+    private readonly TableClient _boardTable;
+    private readonly TableClient _columnTable;
+    private readonly TableClient _swimlaneTable;
+    private readonly TableClient _cardTable;
+    private readonly TableClient _tagTable;
+
+    public BoardRepository (IOptions<CosmosOptions> cosmosOptions)
+    {
+        _tableServiceClient = new TableServiceClient (cosmosOptions.Value.HonuBoards);
+        _boardTable = _tableServiceClient.GetTableClient (tableName: boards);
+        _columnTable = _tableServiceClient.GetTableClient (tableName: columns);
+        _swimlaneTable = _tableServiceClient.GetTableClient (tableName: swimlanes);
+        _cardTable = _tableServiceClient.GetTableClient (tableName: cards);
+        _tagTable = _tableServiceClient.GetTableClient (tableName: tags);
+    }
+
+
+}
